crud, error backpropagation. bannerDB model and converter

+++ internal/domain/request/user.go
@@ -2,7 +2,7 @@ package request
 
 type GetUserRequest struct {
 	Token        string `header:"token"`
-	FeatureId    int    `query:"feature_id"`
-	TagId        int    `query:"tag_id"`
+	FeatureId    int64  `query:"feature_id"`
+	TagId        int64  `query:"tag_id"`
 	LastRevision bool   `query:"use_last_revision"`
 }
